translation-pipeline/requirements-dev.txt
-r requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
moto>=4.2.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0
